feat(ourlogs): Add otel log to sentry log transform (#5128)
#skip-changelog

As per https://github.com/getsentry/relay/issues/5111, we'd like to add
an OTLP logs endpoint to Relay. More details on the motivation can be
found in this internal notion doc:
https://www.notion.so/sentry/Logs-OTLP-Endpoint-2638b10e4b5d8050a147d0dd46df1915
(contains screenshots so can't put them in GH).

Before we implement the endpoint itself, this PR adds a
`otel_to_sentry_log` to the `relay-ourlogs` crate. `otel_to_sentry_log`
is a cleaned up version of the otel log to sentry log transform we had
previously implemented and then removed in
https://github.com/getsentry/relay/pull/5082.

To make the `otel_to_sentry_log`, I extracted out a `otel_value_to_attr`
from the `relay-spans` crate into a `relay-otel` crate. Not sure if this
is the correct level of abstraction, would appreciate feedback.

Skipping changelog as this does not expose any user facing changes, only
new methods and refactors. Once I add the endpoint in a future PR, I
will add a changelog there.