feat(ourlogs): Handle empty span id and trace id in otel log transform (#5150)
resolves
https://linear.app/getsentry/issue/LOGS-352/update-otlp-logs-endpoint-allow-for-optional-tracespan-ids

In the OTEL schema, `span_id` and `trace_id` are optional. In the sentry
logs schema (`ourlog`) `span_id` is also optional, so we can forward
along an empty `span_id`, but `trace_id` is not.

To handle the case where we don't get a `trace_id` from an OTEL log, we
have two approaches we can take:
1. If an OTEL log does not have a `trace_id`, we reject the item
2. We generate a fake `trace_id` and add it to the transformed sentry
log.

We elected to go with 2 because this reduces user confusion as much as
possible as users are often comparing data between two providers when
using OTLP endpoints, and dropping events reduces confidence users have
in Sentry.

Alongside generating a trace id, we also indicate that the trace id was
substituted via `RemarkType::Substituted`. This can be used to fill in
`sentry.meta.X` fields or similar later down the line. Not sure if this
was the best approach, just the one that I took, feedback is welcome if
there is something better that can be done.

I also added some tests.