ref(processing): Adds nothing output (#5196)
Currently not in use, an initial implementation of the sessions
processing needed it, but it was no longer necessary for that.

But in general a `Nothing` output makes sense and I expect it to likely
require it in the future for items which are converted from A->B, like
NEL or other browser reports which will conditionally convert to other
processor inputs (by-products) but will not have an output itself.