fix(ci): Nightly publish for ghcr (#5223)
Looks like composite action for building and publishing images is not
tagging nightly even when `tag_nightly` is set to true.
https://github.com/getsentry/relay/blob/62787c1b1db6063e49b51f7f696e58ef2fedb9a7/.github/workflows/ci.yml#L480

This picks up
https://github.com/getsentry/action-build-and-push-images/pull/19 which
uses the very similar logic as tagging latest in an attempt to fix this
issue. Note that tagging nightly/latest are restricted to workflows on
the default branch, so this cannot be tested on a PR

#skip-changelog