docs(server): Document desired SDK rate limiting behavior per data category (#5188)
So far, [dev
docs](https://develop.sentry.dev/sdk/expected-features/rate-limiting/#definitions)
have been used to inform SDK developers on which data categories to
apply to which envelope items. It has been brought up that this list is
not always complete, and that developers sometimes check the
`DataCategory` definition in Relay for a more complete picture, since it
is the source of truth of data categories used across Sentry. This
however raises new questions, as it is not always clear what relevance
the categories have per SDK.

Add a line to each doc comment for clarification.

INGEST-573.