feat(ourlog): Add vercel log drain transform (#5209)
ref
https://linear.app/getsentry/issue/LOGS-389/add-vercel-log-drain-endpoint-to-relay

This PR updates the `relay-ourlogs` create to add a `VercelLog` struct
and a `vercel_log_to_sentry_log` transform method.
`vercel_log_to_sentry_log` turns an incoming `VercelLog` json payload
structure into an instance of Sentry's `OurLog`.

I haven't hooked up the `vercel_log_to_sentry_log` function to an
endpoint and the rest of the processing pipeline on purpose, I want to
take our time to make sure we are all good with the transform taking
place.

Many of these attributes are new, I've opened a PR in the conventions
repo to document them accordingly:
https://github.com/getsentry/sentry-conventions/pull/163. See this
notion doc for details about the schema and the transform:
https://www.notion.so/sentry/Vercel-Log-Drain-2808b10e4b5d808c95f3ebcfe2b8828a.
The attribute keys here align in naming as per
https://vercel.com/docs/drains/reference/logs.

After this PR merges in, I'll work on exposing this in a new
integrations endpoint, gated behind the feature flag introduced in
https://github.com/getsentry/relay/pull/5208