feat(dev): Add devenv, uv, pre-commit (#5070)
This introduces devenv, uv and pre-commit in accordance with these repo
standards:

https://www.notion.so/sentry/Standard-Spec-devenv-2238b10e4b5d8072a992d8b67d94be22

https://www.notion.so/sentry/Standard-Spec-python-uv-2248b10e4b5d8045b8fff30f8b8b67ca

https://www.notion.so/sentry/Standard-Spec-pre-commit-2238b10e4b5d8019aa1fe1158e39c38e

The ability to run devenv sync and get a dev environment in any repo is
the DX i'm aiming for,
and these standards provide a foundation from which later we can easily
push updates to all repos.

You should be able to checkout this branch, run direnv allow and devenv
sync (might be prompted to run ~/.local/share/sentry-devenv/bin/devenv
update first), which will set up everything needed for development.

Thanks to uv and our internal pypi, setting up the venv is near instant
now (excluding the rust build of `py/` of course).

---

Originally I was going to split this into 2 PRs, where the latter would
configure py/ for uv, but it's not very good DX since uv sync will erase
it as it doesn't know about it.