ref(server): `Item::quantities` now includes the indexed category (#5233)
Mostly for consistency reasons.

An item represents both items, this is also how it is coded in the new
processing pipeline(s). The derived `Counted` impl for `Item` uses
`Item::quantities` and instead of changing the meaning there and having
two different `quantities` implementations, align the method impl and
change usages.