ref(relay): Reduce usage of anyhow (#5242)
Anyhow is pinned to an old version to not capture backtraces. This is a
step in the direction of unlocking the dependency again and still being
able to capture backtraces.

Reduces the usage of `anyhow` especially in processing heavy code, where
capturing backtraces is really expensive and often not even
necessary/used.

`normalize_json` was previously in `relay-dynamic-config`, but only used
in ffi, moved it there and removed the `anyhow` return type.