feat(ai): Add missing gen_ai fields (#5246)
Adds a couple of missing gen_ai fields, sentry convetions PR will be
done as a follow up.

Also marked all of the fields that report token counts as `pii = maybe`
just in case we change scrubbing algorithm, we don't want to filter out
those values.

Closes [TET-1272: Add fields containing `token` to the scrubbing
exceptions in
relay](https://linear.app/getsentry/issue/TET-1272/add-fields-containing-token-to-the-scrubbing-exceptions-in-relay)