fix(spansv2): Correctly map the span name from v1 to v2 (#5253)
The span name was only mapped but not inferred correctly (like we do for
standalone spans).

Also the span name was left in the attributes, when it should've been
removed, as it is redundant.

The segment processor in Sentry, then maps the top level `name` field
again correctly to a `sentry.name` attribute.