ref(derive): Allow usage of ProcessValue macro in other crates (#5256)
Use an absolute crate reference `::relay_event_schema::` instead of
`crate::`, to make it work in the event schema crate, expose `self` as
`relay_event_schema`.

Also changes `EnumSet` to a re-exported type in `::relay_event_schema`
to not have to depend on it directly.

Enables a test that previously was commented out but because it couldn't
be compiled, which now compiles.