feat(tracemetrics): Allow filtering co-occurring keys
Trace metrics has a blocker of having the available group-by and
autocomplete keys being limited to the metric name selected instead of
every attribute key ever ingested. Due to
how the trace item filter rpc (and eap tables) are currently setup, this
is only currently possible with an 'AND' on the key presence, so this is
a workaround adding two temporary meta attributes that will let us
experiment with this behaviour before open beta.