ref(schema): Enforce required attributes for EAP item types (#5258)
Enforces the `required` annotation on fields on EAP items.

Implements a new mode in the schema processor to delete the parent when
a required field fails validation, instead of just deleting the value.

Enables this new option for EAP based items.

Longterm we can think about enabling it for Errors/Transactions as well.