fix(spansv2): Add DSC transaction name to count per root metric (#5278)
The tag was missing for spans v2, there is still a bug in span v1
processing where the local instead of DSC transaction name is added,
which I will fix in a follow-up.

Currently the implementation looks up the DSC multiple times, even in
places where it was already validated to be present, this is not a big
deal, but I plan to look into this in INGEST-609, if we can more
strictly type the DSC (and the presence of the DSC).