feat(relay): Simplify Proxy Mode (#5165)
This is a big change, however this should not break anything. Users that
run Relay in Proxy mode should not experience any negative side effects.

With the recent changes to the Relay modes
(https://github.com/getsentry/relay/pull/5108,
https://github.com/getsentry/relay/pull/5057,
https://github.com/getsentry/relay/pull/5053) it became apparent that
the proxy mode is currently doing more than strictly necessary. That is,
the logic is currently more complex than it needs to be and as such it
also is less efficient than it could be. This PR tries to rectify this
by making Proxy mode more of a 'pure' Proxy by introducing
the `ProxyProcessorService` which is a simplified version of
the `EnvelopeProcessorService`. There are also some changes to the
`HealthCheckService`, `AutoscalingMetricService` and `RelayStats` to
make these compatible.

Fixes: https://linear.app/getsentry/issue/INGEST-581/simplify-proxy-mode