ref(server): Cleanup proxy handling in processing (#5283)
Now that #5165 has landed, we can cleanup the proxy mode handling in the
new processing pipelines.

This simplifies a bunch of code as the output now is always processed.
This was always a bit messy and temporary and put in as a reminder to
deal with proxy mode properly. Now that it is in, we can clean it up :)