ref(server): Cleans up ClientName and PlatformTag, adds Relay client name (#5279)
Cleans up the mess where `ClientName` and `PlatformTag` were added to
the `statsd` module. `ClientName` is closer to `RequestMeta`. I decided
to add a new `meta` module to `envelope` with the longterm idea of
possibly moving the `RequestMeta` types there as well (an extractor
shouldn't own this envelope fundamental type).

For `PlatformTag`, there is already a `statsd/utils` module.

I also added `Relay` to `ClientName`, which is the actual change I
wanted to make.

Oh and there was a bug in the metrics creation, 3 metrics in a row
emitted, the first one used the low cardinality platform, the others
don't.