fix(ds): Tag span count per root with DSC transaction instead of local (#5281)
Closes: TET-1311

The count per root for transactions is tagged with the transaction name
of the trace root (from the DSC). The metric for spans was always tagged
with the local transaction (not from the root).

It is possible that at some point that was intentional. For our purposes
and from my research it seems like a bug and we also want to tag the
span count per root metric with the trace root's transaction name.

Added a `ExtractMetricsConfig` struct, just to have some kind of named
arguments.