feat(spansv2): Reject envelopes with misconfigured DSCs (#5276)
As discussed, reject entire envelopes which contain `SpanV2` item
containers but are missing a DSC, also reject envelopes which do have a
DSC but the contained spans do not conform with the specified DSC.

The intention of this change is to make sure SDKs implement the DSC
correctly.

I ended up using `invalid_dsc` as an outcome reason instead of
`mismatch` (the more specific reason in the processor) because it
matches the existing schema of outcomes.