ref: Remove some methods from `Attributes`
This removes some methods from the `Attributes` struct that would only
immediately delegate to the corresponding method on the underlying
`BTreeMap`. The `BTreeMap` is publicly accessible anyway, so it's not
like we're hiding implementation details.

I've left `contains_key` and `remove` in place for now because it's
kind of convenient for them to exist on `Attributes`, but the the same
applies to them in principle.