ref(redis): Tag connections with client name
This adds a client name to Redis connection pools and connections, which
can then be used to tag Redis metrics.

If the Redis configuration is split between the different use cases
(project configs/cardinality/quotas), each client will have its own name.
If the configuration is unified, they will all have the same name. We
might revisit this in a future PR.