feat(eap): Add trimming processor for EAP items
This adds a new trimming processor specifically for EAP items. The reason
we can't just reuse the existing processor is the size calculation logic.
The existing processor calculates the sizes of values using a special
serializer. For EAP items, specfically their attributes, we want to use
the attribute size logic defined in `eap::size`.

Attributes are processed in order from smallest to greatest total (key
+ value) size. The effect of this is that small attributes are more
likely to be preserved. Attribute keys are taken into account for the
size, but never trimmed—if a key is too long the attribute is just
removed.